Outsourced API documentation creation to new module
